+++ lib/generators/rodauth/templates/app/misc/rodauth_main.rb
@@ -40,6 +40,11 @@ class RodauthMain < Rodauth::Rails::Auth
     # Store password hash in a column instead of a separate table.
     account_password_hash_column :password_hash
 
+    # Passwords shorter than 8 characters are considered weak according to OWASP.
+    password_minimum_length 8
+    # bcrypt has a maximum input length of 72 bytes, truncating any extra bytes.
+    password_maximum_bytes 72
+
     # Set password when creating account instead of when verifying.
     verify_account_set_password? false